+++ src/detectors/ScanningStrategyDetector.spec.ts
@@ -1,3 +1,4 @@
+import nock from 'nock';
 import git from 'simple-git/promise';
 import { createTestContainer } from '../inversify.config';
 import { GitHubNock } from '../test/helpers/gitHubNock';
@@ -6,6 +7,11 @@ jest.mock('simple-git/promise');
 
 describe('ScanningStrategyDetector', () => {
   const mockedGit = <jest.Mock>git;
+
+  beforeEach(() => {
+    nock.cleanAll();
+    nock.enableNetConnect();
+  });
 
   describe('#detect', () => {
     it('local path without remote', async () => {
@@ -23,6 +29,7 @@ describe('ScanningStrategyDetector', () => {
         accessType: undefined,
         localPath: '/local/path',
         remoteUrl: undefined,
+        isOnline: false,
         serviceType: ServiceType.git,
       });
     });
@@ -45,6 +52,31 @@ describe('ScanningStrategyDetector', () => {
         accessType: AccessType.public,
         localPath: '/local/path',
         remoteUrl: repoPath,
+        isOnline: true,
+        serviceType: ServiceType.github,
+      });
+    });
+
+    it('offline local path with remote public GitHub and auth', async () => {
+      const repoPath = 'git@github.com:DXHeroes/dx-scanner.git';
+      nock.disableNetConnect();
+      new GitHubNock('1', 'DXHeroes', 1, 'dx-scanner').getRepo('').reply(500);
+
+      mockedGit.mockImplementation(() => {
+        return {
+          checkIsRepo: () => true,
+          getRemotes: () => [{ name: 'origin', refs: { fetch: repoPath, push: repoPath } }],
+        };
+      });
+      const container = createTestContainer({ uri: '/local/path', auth: 'fake_token' });
+
+      const result = await container.scanningStrategyDetector.detect();
+
+      expect(result).toEqual({
+        accessType: AccessType.unknown,
+        localPath: '/local/path',
+        remoteUrl: repoPath,
+        isOnline: false,
         serviceType: ServiceType.github,
       });
     });
@@ -67,6 +99,31 @@ describe('ScanningStrategyDetector', () => {
         accessType: 'public',
         localPath: '/local/path',
         remoteUrl: repoPath,
+        isOnline: true,
+        serviceType: ServiceType.github,
+      });
+    });
+
+    it('offline local path with remote public GitHub and no auth', async () => {
+      const repoPath = 'git@github.com:DXHeroes/dx-scanner.git';
+      nock.disableNetConnect();
+      new GitHubNock('1', 'DXHeroes', 1, 'dx-scanner').getRepo('').reply(500);
+
+      mockedGit.mockImplementation(() => {
+        return {
+          checkIsRepo: () => true,
+          getRemotes: () => [{ name: 'origin', refs: { fetch: repoPath, push: repoPath } }],
+        };
+      });
+      const container = createTestContainer({ uri: '/local/path' });
+
+      const result = await container.scanningStrategyDetector.detect();
+
+      expect(result).toEqual({
+        accessType: AccessType.unknown,
+        localPath: '/local/path',
+        remoteUrl: repoPath,
+        isOnline: false,
         serviceType: ServiceType.github,
       });
     });
@@ -89,6 +146,31 @@ describe('ScanningStrategyDetector', () => {
         accessType: AccessType.unknown,
         localPath: '/local/path',
         remoteUrl: repoPath,
+        isOnline: true,
+        serviceType: ServiceType.github,
+      });
+    });
+
+    it('offline local path with remote private GitHub', async () => {
+      const repoPath = 'git@github.com:DXHeroes/dx-scanner-private.git';
+      nock.disableNetConnect();
+      new GitHubNock('1', 'DXHeroes', 1, 'dx-scanner-private').getRepo('').reply(500);
+
+      mockedGit.mockImplementation(() => {
+        return {
+          checkIsRepo: () => true,
+          getRemotes: () => [{ name: 'origin', refs: { fetch: repoPath, push: repoPath } }],
+        };
+      });
+
+      const container = createTestContainer({ uri: '/local/path', auth: 'bad AT' });
+      const result = await container.scanningStrategyDetector.detect();
+
+      expect(result).toEqual({
+        accessType: AccessType.unknown,
+        localPath: '/local/path',
+        remoteUrl: repoPath,
+        isOnline: false,
         serviceType: ServiceType.github,
       });
     });
@@ -104,6 +186,7 @@ describe('ScanningStrategyDetector', () => {
         accessType: AccessType.public,
         localPath: undefined,
         remoteUrl: repoPath,
+        isOnline: true,
         serviceType: ServiceType.github,
       });
     });
@@ -118,6 +201,7 @@ describe('ScanningStrategyDetector', () => {
         accessType: AccessType.unknown,
         localPath: undefined,
         remoteUrl: repoPath,
+        isOnline: true,
         serviceType: ServiceType.github,
       });
     });
@@ -133,6 +217,7 @@ describe('ScanningStrategyDetector', () => {
         accessType: AccessType.public,
         localPath: undefined,
         remoteUrl: `https://${repoPath}`,
+        isOnline: true,
         serviceType: ServiceType.github,
       });
     });